Archive/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot-nl-sql"
version = "0.1.0"
description = "Natural Language to SQL Chatbot using Ollama and FastAPI"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]

dependencies = [
    "fastapi==0.99.1",
    "uvicorn[standard]==0.23.2",
    "httpx==0.24.1",
    "asyncpg==0.31.0",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/example/chatbot-nl-sql"
Repository = "https://github.com/example/chatbot-nl-sql.git"

[tool.uv]
managed = true

[tool.uv.pip]
python-requires = ">=3.11,<3.15"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
